kolemsi_bot/requirements.txt
arrow
pyautogui
Pillow
python-dotenv
telebot
requests
beautifulsoup4